+++ packages/components/components/elvis-dropdown/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Elvia Dropdown
 
+## 2.0.2 (30.11.21)
+
+### Patches
+
+- Now using elvis-colors package for colors.
+- Now using elvis-typography package for typography.
+
 ## 2.0.1 (30.11.21)
 
 - Web component wrapper fix