+++ packages/components/components/elvis-modal/CHANGELOG.md
@@ -1,6 +1,14 @@
 # Elvia Modal Changelog
 
+## 1.3.3 (30.11.21)
+
+### Patch
+
+- Now using new version of elvis-colors.
+
 ## 1.3.2 (30.11.21)
+
+### Patches
 
 - Web component wrapper fix